move data cleaning code to new module
