+++ src/vivarium_helpers/vph_output/measures.py
@@ -1,6 +1,7 @@
 import pandas as pd
 from .loading import VPHOutput
 from .operations import VPHOperator, list_columns
+from .cleaning import clean_vph_output
 
 class VPHResults(VPHOutput):
 
@@ -30,17 +31,7 @@ class VPHResults(VPHOutput):
 
     def _clean_vph_output(self):
         """Reformat transformed count data to make more sense."""
-        # TODO: Maybe implement this as a module method and
-        # call it from here, since it will probably be a lot of code
-        #
-        # # Make the wasting and disease transition count dataframes better
-        # clean_data.update(
-        #     {table_name: clean_transition_df(table)
-        #      for table_name, table in data.items()
-        #      if table_name.endswith('transition_count')})
-        if 'ylds' in self and 'cause_of_disability' in self['ylds']:
-            self['ylds'].rename(
-                        columns={'cause_of_disability': 'cause'}, inplace=True)
+        clean_vph_output(self)
 
     def compute_dalys(self):
         # TODO: Handle the case where one of YLLs or YLDs
@@ -248,35 +239,3 @@ class VPHResults(VPHOutput):
 
 ########################
 #### Module methods ####
-
-def split_measure_and_transition_columns(transition_df):
-    """Separates the transition from the measure in the strings in the 'measure'
-    columns in a transition count dataframe, and puts these in separate 'transition'
-    and 'measure' columns.
-    """
-    return (transition_df
-            .assign(transition=lambda df: df['measure']
-                    .str.replace('_event_count', '')) # Older models label this event_count
-            .assign(measure='transition_count') # Name the measure 'transition_count' rather than 'event_count'
-           )
-
-def extract_transition_states(transition_df):
-    """Gets the 'from state' and 'to state' from the transitions in a transition count dataframe,
-    after the transition has been put in its own 'transition' column by the `split_measure_and_transition_columns`
-    function.
-    """
-    states_from_transition_pattern = r"^(?P<from_state>\w+)_to_(?P<to_state>\w+)$"
-    # Renaming the 'susceptible_to' states is a hack to deal with the fact there's not a unique string
-    # separating the 'from' and 'to' states -- it should be '__to__' instead of '_to_' or something
-    states_df = (
-        transition_df['transition']
-        .str.replace("susceptible_to", "without") # Remove word 'to' from all states so we can split transitions on '_to_'
-        .str.extract(states_from_transition_pattern) # Create dataframe with 'from_state' and 'to_state' columns
-        .apply(lambda col: col.str.replace("without", "susceptible_to")) # Restore original state names
-    )
-    return states_df
-
-# Define a function to make the transition count dataframes better
-def clean_transition_df(df):
-    df = split_measure_and_transition_columns(df)
-    return df.join(extract_transition_states(df))
